functionality of inputs works
