+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
   "eslint.validate": ["typescript", "javascript"],
   "files.eol": "\n",
-  "cSpell.words": ["Elems"]
+  "cSpell.words": ["Almarai", "Elems"]
 }
